+++ tests/LumenModuleTest.php
@@ -117,8 +117,6 @@ class LumenModuleTest extends BaseTestCase
     public function it_checks_module_enabled_status()
     {
         $this->assertTrue($this->module->enabled());
-        $this->assertTrue($this->module->active());
-        $this->assertFalse($this->module->notActive());
         $this->assertFalse($this->module->disabled());
     }